backend/requirements.txt
# Aquaforest RAG System Dependencies

# Core dependencies
openai>=1.35.0
pinecone[grpc]>=7.0.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-hyperbrowser>=0.1.0
pymupdf>=1.24.0
pydantic>=2.7.0
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Utils
tqdm>=4.66.0
numpy>=1.26.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Development dependencies
pytest>=8.2.0
pytest-asyncio>=0.23.0
black>=24.4.0
flake8>=7.0.0

# FastAPI server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
slowapi>=0.1.9

# Optional for better performance (Linux/Mac only)
uvloop>=0.19.0; sys_platform != "win32"